Fix compilation errors post merge
we still have few undefined references, which need to be solved.

We can do the following:
Sound, Video/msm can be migrated completely towards CAF, instead of maintaining
a merge of Xiaomi and CAF. We can take it up when we get time or we have issues
with merge .. Err.